Kubernetes Launch Script Functionality Improvement
Change the way this script handles being called upon so that it can
support more than just the build-setup.sh. It will now call a
templated YAML file to for the kubectl to launch. This template will
be filled out and default values for the template will be set in a
case check that looks at the invoker and determines what variables
are needed to launch the YAML file. To satisfy these changes updated
the build-setup.sh, and added a templates folder.

+++ kubernetes/kubernetes-launch.sh
@@ -2,8 +2,8 @@
 ###############################################################################
 #
 # Script used to assist in launching Kubernetes jobs/pods. Expects to be used
-# as an supplemental script to the build-setup.sh script as such will use some
-# of the variables it expects to carry over from that script.
+# as an supplemental script to the scripts that want to launch their containers
+# on a Kubernetes cluster.
 #
 ###############################################################################
 #
@@ -21,9 +21,6 @@
 ###############################################################################
 # Variables used to create Kubernetes Job:
 #  namespace    = the namespace to be used within the Kubernetes cluster
-#  hclaim       = name of the Jenkins slave home PVC on Kubernetes cluster
-#  sclaim       = name of the shared state cache PVC on Kubernetes cluster
-#  oclaim       = name of OpenBMC cache repo PVC on the Kubernetes cluster
 #  registry     = the registry to use to pull and push images
 #  imgplsec     = the image pull secret used to access registry if needed
 #  timeout      = the amount of time in seconds that the build will wait for
@@ -36,16 +33,39 @@
 
 # Kubernetes Variables
 namespace=${namespace:-openbmc}
-hclaim=${hclaim:-jenkins}
-sclaim=${sclaim:-shared-state-cache}
-oclaim=${oclaim:-openbmc-reference-repo}
 imgrepo=${imgrepo:-master.cfc:8500/openbmc/}
 imgplsec=${imgplsec:-regkey}
 timeout=${timeout:-60}
 
-# Give the Docker image and the pod a name
-imgname=${imgname:-${imgrepo}${distro}:${imgtag}-${ARCH}}
-podname=${podname:-openbmc${BUILD_ID}-${target}-builder}
+# Options which decide script behavior
+invoker=${invoker:-${1}}
+log=${log:-${2}}
+purge=${purge:-${3}}
+launch=${launch:-${4}}
+
+# Set the variables for the specific invoker to fill in the YAML template
+case ${invoker} in
+  OpenBMC-build)
+    # Should be invoked and sourced in the build-setup.sh
+    hclaim=${hclaim:-jenkins}
+    sclaim=${sclaim:-shared-state-cache}
+    oclaim=${oclaim:-openbmc-reference-repo}
+    imgname=${imgname:-${imgrepo}${distro}:${imgtag}-${ARCH}}
+    podname=${podname:-openbmc${BUILD_ID}-${target}-builder}
+    ;;
+  QEMU-build)
+    ;;
+  QEMU-launch)
+    ;;
+  XCAT-launch)
+    ;;
+  generic)
+    ;;
+  *)
+    exit 1
+    ;;
+esac
+
 
 # Build the Docker image, using the Dockerfile carried from build-setup.sh
 docker build -t ${imgname} - <<< "${Dockerfile}"
@@ -53,116 +73,10 @@ docker build -t ${imgname} - <<< "${Dockerfile}"
 # Push the image that was built to the image repository
 docker push ${imgname}
 
-if [[ "${launch}" == "pod" ]]; then
-  yamlfile=$(cat << EOF
-  apiVersion: v1
-  kind: Pod
-  metadata:
-    name: ${podname}
-    namespace: ${namespace}
-  spec:
-    nodeSelector:
-      worker: "true"
-      arch: ${ARCH}
-    volumes:
-    - name: home
-      persistentVolumeClaim:
-        claimName: ${hclaim}
-    - name: sscdir
-      persistentVolumeClaim:
-        claimName: ${sclaim}
-    - name: obmccache
-      persistentVolumeClaim:
-        claimName: ${oclaim}
-    hostNetwork: True
-    containers:
-    - image: ${imgname}
-      name: builder
-      command: ["${WORKSPACE}/build.sh"]
-      workingDir: ${HOME}
-      env:
-      - name: WORKSPACE
-        value: ${WORKSPACE}
-      - name: obmcdir
-        value: ${obmcdir}
-      securityContext:
-        capabilities:
-          add:
-          - SYS_ADMIN
-      volumeMounts:
-      - name: home
-        mountPath: ${HOME}
-      - name: sscdir
-        mountPath: ${sscdir}
-      - name: obmccache
-        mountPath: ${ocache}
-    imagePullSecrets:
-    - name: ${imgplsec}
-EOF
-)
-
-elif [[ "${launch}" == "job" ]]; then
-  yamlfile=$(cat << EOF
-  apiVersion: batch/v1
-  kind: Job
-  metadata:
-    name: openbmc${BUILD_ID}-${target}
-    namespace: ${namespace}
-    labels:
-      app: openbmc
-      stage: build
-  spec:
-    template:
-      metadata:
-        name: ${podname}
-        labels:
-          target: ${target}
-      spec:
-        nodeSelector:
-          worker: "true"
-          arch: ${ARCH}
-        volumes:
-        - name: home
-          persistentVolumeClaim:
-            claimName: ${hclaim}
-        - name: sscdir
-          persistentVolumeClaim:
-            claimName: ${sclaim}
-        - name: obmccache
-          persistentVolumeClaim:
-            claimName: ${oclaim}
-        restartPolicy: Never
-        hostNetwork: True
-        containers:
-        - image: ${imgname}
-          name: builder
-          command: ["${WORKSPACE}/build.sh"]
-          workingDir: ${HOME}
-          env:
-          - name: WORKSPACE
-            value: ${WORKSPACE}
-          - name: obmcdir
-            value: ${obmcdir}
-          securityContext:
-            capabilities:
-              add:
-              - SYS_ADMIN
-          volumeMounts:
-          - name: home
-            mountPath: ${HOME}
-          - name: sscdir
-            mountPath: ${sscdir}
-          - name: obmccache
-            mountPath: ${ocache}
-        imagePullSecrets:
-        - name: ${imgplsec}
-EOF
-)
-fi
-
+yamlfile=$(eval "echo \"$(<./Templates/${invoker}-${launch}.yaml)\"" )
 kubectl create -f - <<< "${yamlfile}"
 
-# Wait for Pod to be running before tailing log file
+# Wait for Pod to be running
 while [ -z "$(kubectl describe pod ${podname} -n ${namespace} | grep Status: | grep Running)" ]; do
   if [ ${timeout} -lt 0 ];then
     kubectl delete -f - <<< "${yamlfile}"
@@ -173,9 +87,14 @@ while [ -z "$(kubectl describe pod ${podname} -n ${namespace} | grep Status: | g
     let timeout-=1
   fi
 done
+echo "Pod is Running"
 
 # Once pod is running track logs
-kubectl logs -f ${podname} -n ${namespace}
+if [[ "${log}" == true ]]; then 
+  kubectl logs -f ${podname} -n ${namespace}
+fi
 
-# When job is completed wipe the job
-kubectl delete -f - <<< "${yamlfile}"
+# Delete the object if purge is true
+if [[ "${purge}" == true ]]; then
+  kubectl delete -f - <<< "${yamlfile}"
+fi
